# tcl-web/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ToneCards Lab API",
    "description": "ToneCards Lab API for managing musical language flashcards",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sharing/decks/{deck_id}/viewers": {
      "get": {
        "tags": [
          "sharing"
        ],
        "summary": "Get Deck Viewers",
        "description": "Get list of users who have viewer access to a deck.",
        "operationId": "get_deck_viewers_api_v1_sharing_decks__deck_id__viewers_get",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ViewersResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sharing"
        ],
        "summary": "Edit Deck Viewers",
        "description": "Add or remove viewers from a deck.",
        "operationId": "edit_deck_viewers_api_v1_sharing_decks__deck_id__viewers_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Owner Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditViewersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ViewersResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/decks/{deck_id}/share-by-email": {
      "post": {
        "tags": [
          "sharing"
        ],
        "summary": "Share Deck By Email",
        "description": "Send an email invitation to share a deck.",
        "operationId": "share_deck_by_email_api_v1_sharing_decks__deck_id__share_by_email_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          },
          {
            "name": "sender_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sender Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareByEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ShareByEmailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/invitations/{invitation_id}": {
      "get": {
        "tags": [
          "sharing"
        ],
        "summary": "Get Invitation",
        "description": "Get invitation details.",
        "operationId": "get_invitation_api_v1_sharing_invitations__invitation_id__get",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ShareByEmailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sharing"
        ],
        "summary": "Revoke Invitation",
        "description": "Revoke a pending invitation.",
        "operationId": "revoke_invitation_api_v1_sharing_invitations__invitation_id__delete",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/invitations/{invitation_id}/accept": {
      "post": {
        "tags": [
          "sharing"
        ],
        "summary": "Accept Invitation",
        "description": "Accept a deck sharing invitation.",
        "operationId": "accept_invitation_api_v1_sharing_invitations__invitation_id__accept_post",
        "parameters": [
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invitation Id"
            }
          },
          {
            "name": "user_email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/invitations/email/{email}": {
      "get": {
        "tags": [
          "sharing"
        ],
        "summary": "Get User Invitations",
        "description": "Get pending invitations for a user's email.",
        "operationId": "get_user_invitations_api_v1_sharing_invitations_email__email__get",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_List_ShareByEmailResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/decks/{deck_id}/viewers/{viewer_id}/remove": {
      "post": {
        "tags": [
          "sharing"
        ],
        "summary": "Remove Viewer",
        "description": "Remove a viewer from a deck.",
        "operationId": "remove_viewer_api_v1_sharing_decks__deck_id__viewers__viewer_id__remove_post",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Deck Id"
            }
          },
          {
            "name": "viewer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Viewer Id"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Owner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse"
      },
...
      "ApiResponse_List_ShareByEmailResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ShareByEmailResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[List[ShareByEmailResponse]]"
      },
...
      "ApiResponse_ShareByEmailResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareByEmailResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[ShareByEmailResponse]"
      },
...
      "ApiResponse_ViewersResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Indicates if the request was successful",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Response message or error description"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ViewersResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata",
            "description": "Response metadata"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Error details if any"
          }
        },
        "type": "object",
        "title": "ApiResponse[ViewersResponse]"
      },
...
      "EditViewersRequest": {
        "properties": {
          "add_viewer_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Add Viewer Ids",
            "description": "User IDs to add as viewers"
          },
          "remove_viewer_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Remove Viewer Ids",
            "description": "User IDs to remove from viewers"
          }
        },
        "type": "object",
        "title": "EditViewersRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Metadata": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Response timestamp"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request identifier"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "API version",
            "default": "1.0"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count",
            "description": "Total number of items (for paginated responses)"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page",
            "description": "Current page number"
          },
          "page_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Size",
            "description": "Number of items per page"
          },
          "execution_time_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Time Ms",
            "description": "Request execution time in milliseconds"
          }
        },
        "type": "object",
        "title": "Metadata"
      },
...
      "ShareByEmailRequest": {
        "properties": {
          "recipient_email": {
            "type": "string",
            "format": "email",
            "title": "Recipient Email",
            "description": "Email address of recipient"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Optional invitation message"
          }
        },
        "type": "object",
        "required": [
          "recipient_email"
        ],
        "title": "ShareByEmailRequest"
      },
      "ShareByEmailResponse": {
        "properties": {
          "invitation_id": {
            "type": "string",
            "title": "Invitation Id",
            "description": "UUID of the invitation"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status of the invitation",
            "default": "sent"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "ISO 8601 timestamp when invitation expires"
          }
        },
        "type": "object",
        "required": [
          "invitation_id"
        ],
        "title": "ShareByEmailResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ViewersResponse": {
        "properties": {
          "deck_id": {
            "type": "string",
            "title": "Deck Id",
            "description": "UUID of the deck"
          },
          "viewer_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Viewer Ids",
            "description": "List of user IDs with viewer access"
          }
        },
        "type": "object",
        "required": [
          "deck_id"
        ],
        "title": "ViewersResponse"
      }
    }
  }
}